Superclass Cryptor: Created Superclass and refactored Encryptor into it. All tests pass
