+++ test/enigma_test.rb
@@ -32,7 +32,6 @@ class EnigmaTest < Minitest::Test
     assert !('' == actual[:encryption])
     assert_instance_of Integer, actual[:key].to_i
     assert_instance_of Integer, actual[:date].to_i
-    binding.pry
   end
 
 end
